Work around bug in OS X pthread_rwlock_init
Calling pthread_rwlock_init with a pthread_rwlock_t that is not
initialized to zeros sometimes causes a segfault. This crashes Oracle's
Java 7 on my machine with a message like this:

  A fatal error has been detected by the Java Runtime Environment:

  SIGSEGV (0xb) at pc=0x00007fff8e6c86ac, pid=78941, tid=4875

  JRE version: 7.0_07-b10
  Java VM: Java HotSpot(TM) 64-Bit Server VM (23.3-b01 mixed mode bsd-amd64 compressed oops)
  Problematic frame:
  C  [libsystem_c.dylib+0x1a6ac]  pthread_rwlock_init+0x50

Even worse, it crashes Apple's Java 6 with the even less obvious:

  Invalid memory access of location 0x7fa000000000 rip=0x7fff8e6c86ac

This bug took a very long time to track down. Thanks to a comment here
for finally pointing me in the right direction:

https://llvm.org/svn/llvm-project/llvm/tags/RELEASE_28/lib/System/RWMutex.cpp